Add core package to handle jobs, streaming output, and storage
This adds JobStore which lets you store jobs and get them by ID.
Job which wraps a command and exposes calls to start, stop, get
the status, and stream the output of a job. Stream which allows
for publishing to many subscibers and will be used to allow
clients to stream a jobs output.